requirements.txt
annotated-types==0.7.0
anyio==4.6.2
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
